Re-implement "hover" functionality
The normal version is `blessed` is decent, but it does not support
nested elements. So for example if you have `hover` styles on a "box",
and there's a "text" element inside and you mouse over the text element,
then the "box" styling gets de-activated which is lame.

This version also retains the "hovered" state while holding the
mouse down (i.e. during dragging, etc.), which the default blessed
implementation does not.